refactor(section): 인수 테스트 단계를 한 파일로 모았다.
